minor changes to glmsingle script (debugging)

+++ emotion_word_glmsingle/evaluate_GLMsingle_emotion_words.py
@@ -111,7 +111,7 @@ def main(raw_args=None):
         pcstop = '-0'  # make sure the string names are correct!
     fracs = args.fracs
     if args.brain_R2 is not None:
-        brain_R2 = int(args.brain_R2)
+        brain_R2 = float(args.brain_R2)
     else:
         brain_R2 = args.brain_R2
